Fix warnings with clang host compiler.

+++ nvbench/detail/measure_cold.cu
@@ -43,8 +43,8 @@ measure_cold_base::measure_cold_base(state &exec_state)
 {
   if (m_min_samples > 0)
   {
-    m_cuda_times.reserve(m_min_samples);
-    m_cpu_times.reserve(m_min_samples);
+    m_cuda_times.reserve(static_cast<std::size_t>(m_min_samples));
+    m_cpu_times.reserve(static_cast<std::size_t>(m_min_samples));
   }
 }
